Most of the transformation is now done.
Most of the transformation is done, however the corner points are still
pending.

Most of the time of this section was spent debuging an valgrind error
where the uniinitialized bits returned by the valgrind program were
causing errors. By using calloc, this problem was solved since the bits
are automatically initialized by the intrinsic function. It is slower
but safer !

Cheers.

 On branch develop
 Changes to be committed:
	new file:   allocation.c
	modified:   main.c
	modified:   output.c
	modified:   structs.h
	modified:   transformation.c