Adding the starting runnig version.
Note my dear. Changing O.S. is a lot of fun, however, doing this while
on a deadline to send the final project is a mess. The cgns library was
not being properly pointed out. So the mesh was not being read properly.
Having a version of the cgns and hdf5 library installed in the system
did not helped as well.

Re-compile the libs and move on...

 On branch develop
 Your branch is up to date with 'origin/develop'.

 Changes to be committed:
	modified:   Makefile
	modified:   input.in
	modified:   jacobian.c
	modified:   main.c
	modified:   prototypes.h
	modified:   rhs_steger_warming.c