Adding the eps on the steger warming.
It is important to note that the splited eigenvalues are not passing the
consistent test in the implicit version of the steger warming.

It is imporant to note that the tests involving the jacobian matrices
shall continue in order to understantd the influence of the jacobians on
the convergence rate.

 On branch develop
 Your branch is up to date with 'origin/develop'.

 Changes to be committed:
	modified:   input.in
	modified:   prototypes.h
	modified:   steger_warming.c